testing/web-platform/mozilla/tests/credential-management/identity.create-store-collect.sub.https.html

<!doctype html>
<head>
<title>Test creation of 'identity' credentials</title>
</head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
'use strict';

const ccs = SpecialPowers.Cc[
  "@mozilla.org/browser/credentialchooserservice;1"
].getService(SpecialPowers.Ci.nsICredentialChooserService);


promise_test(async t => {
  let initData = {
    id : "id1",
    token: "token1",
  };

  let result1 = await navigator.credentials.create({
    identity : initData,
  });
  assert_not_equals(result1, undefined, "A credential is returned from the CredentialContainer.");
  assert_equals(result1.type, "identity", "The credential is an IdentityCredential");
  assert_equals(result1.id, "id1", "The identity matches the argument");
  assert_equals(result1.token, "token1", "The token matches the argument");


  let result2 = new IdentityCredential(initData);
  assert_not_equals(result2, undefined, "A credential is returned from new.");
  assert_equals(result2.type, "identity", "The credential is an IdentityCredential");
  assert_equals(result2.id, "id1", "The identity matches the argument");
  assert_equals(result2.token, "token1", "The token matches the argument");
}, "Basic valid credential creation");


promise_test(async t => {
  let initData = {
    id_missing : "invalid"
  };


  await assert_throws_js(TypeError, () => new IdentityCredential(initData), "Create with missing id field in identity credential constructor throws TypeError.");
  return promise_rejects_js(t, TypeError, navigator.credentials.create({
    identity : initData,
  }), "Create with missing id field in identity credential creation throws TypeError.");

}, "Basic invalid credential creation");

promise_test(async t => {
  let initData = {
    id : "id1"
  };

  let result1 = await navigator.credentials.create({
    identity : initData,
  });

  let result2 = await navigator.credentials.store(result1);
  assert_equals(result2, undefined, "The result of a store should be a promise that resolves to undefined.")

  let result3 = await navigator.credentials.store(result1);
  assert_equals(result3, undefined, "The result of a duplicate store should be a promise that resolves to undefined.")

}, "Basic identity credential store suceeds");


promise_test(async t => {
  let initData = {
    id : "id1",
  };

  let result1 = await navigator.credentials.create({
    identity : initData,
  });
  await navigator.credentials.store(result1);

  await ccs.testMakeChoice(SpecialPowers.wrap(window).browsingContext, "id1");
  let promise2 = navigator.credentials.get({identity: {providers: [{origin: window.origin}]}});
  assert_true(promise2 instanceof Promise, "The return value of get must be a promise.");
  let result2 = await promise2;
  assert_equals(result2, null, "A credential is not returned from the CredentialContainer when the RP is not on the allowlist.");

  let initDataWithOrigin = {
    id : "id1",
    effectiveOrigins : [window.origin],
  };

  let credential = await navigator.credentials.create({
    identity : initDataWithOrigin,
  });
  await navigator.credentials.store(credential);

  await ccs.testMakeChoice(SpecialPowers.wrap(window).browsingContext, "id1");
  let getCredentialPromise = navigator.credentials.get({identity: {providers: [{origin: window.origin}]}});
  let credentialGotten = await getCredentialPromise;
  assert_not_equals(credentialGotten, undefined, "A credential is returned from the CredentialContainer.");
  assert_equals(credentialGotten.type, "identity", "The credential is an IdentityCredential");
  assert_equals(credentialGotten.id, "id1", "The identity matches the argument");


  let promise3 = navigator.credentials.get({identity: {providers: []}});
  assert_true(promise3 instanceof Promise, "The return value of get must be a promise.");
  let result3 = await promise3;
  assert_equals(result3, null, "A credential is not returned from the CredentialContainer when no providers are given.");


  let promise4 = navigator.credentials.get({identity: {providers: [{origin: "https://{{hosts[alt][]}}:{{ports[https][0]}}"}]}});
  assert_true(promise4 instanceof Promise, "The return value of get must be a promise.");
  let result4 = await promise4;
  assert_equals(result4, null, "A credential is not returned from the CredentialContainer when a provider with no credentials is given.");
}, "Basic same-origin identity credential collect suceeds");

promise_test(async t => {

  let initDataWithOrigin = {
    id : "id1",
    effectiveQueryURL : `https://{{hosts[][]}}:{{ports[https][0]}}/_mozilla/credential-management/support/cors.py`,
  };

  let credential = await navigator.credentials.create({
    identity : initDataWithOrigin,
  });
  await navigator.credentials.store(credential);

  let promise2 = navigator.credentials.get({identity: {providers: [{origin: window.origin}]}});
  assert_true(promise2 instanceof Promise, "The return value of get must be a promise.");
  let result2 = await promise2;
  assert_equals(result2, null, "A credential is not returned from the CredentialContainer when the effectiveQueryURL is not provided.");


  await ccs.testMakeChoice(SpecialPowers.wrap(window).browsingContext, "id1");
  let getCredentialPromise = navigator.credentials.get({identity: {providers: [{origin: window.origin, effectiveQueryURL: `https://{{hosts[][]}}:{{ports[https][0]}}/_mozilla/credential-management/support/cors.py`}]}});
  assert_true(getCredentialPromise instanceof Promise, "The return value of get must be a promise.");
  let credentialGotten = await getCredentialPromise;
  assert_not_equals(credentialGotten, undefined, "A credential is returned from the CredentialContainer.");
  assert_equals(credentialGotten.type, "identity", "The credential is an IdentityCredential");
  assert_equals(credentialGotten.id, "id1", "The identity matches the argument");

}, "Basic same-origin identity credential collect with effectiveQueryURL suceeds");

promise_test(async t => {
  let initDataWithOrigin = {
    id : "id1",
    effectiveQueryURL : `https://{{hosts[][]}}:{{ports[https][0]}}/_mozilla/credential-management/support/no-cors.py`,
  };

  let credential = await navigator.credentials.create({
    identity : initDataWithOrigin,
  });
  await navigator.credentials.store(credential);

  let promise2 = navigator.credentials.get({identity: {providers: [{origin: window.origin}]}});
  assert_true(promise2 instanceof Promise, "The return value of get must be a promise.");
  let result2 = await promise2;
  assert_equals(result2, null, "A credential is not returned from the CredentialContainer when the effectiveQueryURL is not provided.");


  let getCredentialPromise = navigator.credentials.get({identity: {providers: [{origin: window.origin, effectiveQueryURL: `https://{{hosts[][]}}:{{ports[https][0]}}/_mozilla/credential-management/support/no-cors.py`}]}});
  let credentialGotten = await getCredentialPromise;
  assert_not_equals(credentialGotten, undefined, "A credential is returned from the CredentialContainer.");
  assert_equals(credentialGotten.type, "identity", "The credential is an IdentityCredential");
  assert_equals(credentialGotten.id, "id1", "The identity matches the argument");

}, "Basic same-origin identity credential collect with effectiveQueryURL and no ACAO works because CORS is not needed");


promise_test(async t => {
  let createdPromise = new Promise((resolve) => {
    window.addEventListener(
      "message",
      (event) => {
        if (event.origin == `https://{{hosts[alt][]}}:{{ports[https][0]}}`) {
          if (event.data == "created") {
            resolve();
          }
        }
      }
    );
  });
  let win = window.open(`https://{{hosts[alt][]}}:{{ports[https][0]}}/_mozilla/credential-management/support/identity.provider.sub.html`);

  await createdPromise;
  win.close();

  let getCredentialPromise, credentialGotten;

  await ccs.testMakeChoice(SpecialPowers.wrap(window).browsingContext, "no-options");
  getCredentialPromise = navigator.credentials.get({identity: {providers: [{origin: "https://{{hosts[alt][]}}:{{ports[https][0]}}"}]}});
  assert_true(getCredentialPromise instanceof Promise, "The return value of get must be a promise.");
  credentialGotten = await getCredentialPromise;
  assert_equals(credentialGotten, null, "A optionless credential is not returned from the CredentialContainer.");

  await ccs.testMakeChoice(SpecialPowers.wrap(window).browsingContext, "allow-alt");
  getCredentialPromise = navigator.credentials.get({identity: {providers: [{origin: "https://{{hosts[alt][]}}:{{ports[https][0]}}"}]}});
  assert_true(getCredentialPromise instanceof Promise, "The return value of get must be a promise.");
  credentialGotten = await getCredentialPromise;
  assert_equals(credentialGotten, null, "A allow-alt credential is not returned from the CredentialContainer.");

  await ccs.testMakeChoice(SpecialPowers.wrap(window).browsingContext, "allow-primary");
  getCredentialPromise = navigator.credentials.get({identity: {providers: [{origin: "https://{{hosts[alt][]}}:{{ports[https][0]}}"}]}});
  assert_true(getCredentialPromise instanceof Promise, "The return value of get must be a promise.");
  credentialGotten = await getCredentialPromise;
  assert_not_equals(credentialGotten, null, "A allow-primary credential is returned from the CredentialContainer.");

  await ccs.testMakeChoice(SpecialPowers.wrap(window).browsingContext, "allow-cors-failure");
  getCredentialPromise = navigator.credentials.get({identity: {providers: [{origin: "https://{{hosts[alt][]}}:{{ports[https][0]}}", effectiveQueryURL: "https://{{hosts[alt][]}}:{{ports[https][0]}}/_mozilla/credential-management/support/no-cors.py"}]}});
  assert_true(getCredentialPromise instanceof Promise, "The return value of get must be a promise.");
  credentialGotten = await getCredentialPromise;
  assert_equals(credentialGotten, null, "A allow-cors-failure credential is not returned from the CredentialContainer.");

  await ccs.testMakeChoice(SpecialPowers.wrap(window).browsingContext, "allow-cors-success");
  getCredentialPromise = navigator.credentials.get({identity: {providers: [{origin: "https://{{hosts[alt][]}}:{{ports[https][0]}}", effectiveQueryURL: "https://{{hosts[alt][]}}:{{ports[https][0]}}/_mozilla/credential-management/support/cors.py"}]}});
  assert_true(getCredentialPromise instanceof Promise, "The return value of get must be a promise.");
  credentialGotten = await getCredentialPromise;
  assert_not_equals(credentialGotten, null, "A allow-cors-success credential is returned from the CredentialContainer.");

}, "Cross-origin identity credential collection is correctly filtered by arguments");

</script>
